Minor Refactoring
Removing problematic packages
