+++ src/test/java/com/sparta/badgerBytes/webTesting/cucumber/stepdefs/InvalidLoginDetails.java
@@ -1,4 +1,4 @@
-package com.sparta.badgerBytes.webTesting.cucumber.stepdefs.InvalidLogin;
+package com.sparta.badgerBytes.webTesting.cucumber.stepdefs;
 
 import com.sparta.badgerBytes.webTesting.cucumber.stepdefs.BackgroundStepdefs;
 import com.sparta.badgerBytes.webTesting.pom.pages.HomePage;
